updating

+++ src/testing.py
@@ -35,7 +35,7 @@ def evaluate(
             for im in img:
                 if isGraphEnc:
                     G = torch.load(f"{img_tnsr_path}/{int(im.item())}.pt")
-                    _features_list.append(G.x)
+                    _features_list.append(G.x.float())
                     _edge_list.append(G.edge_index) 
                 
                 if isVitEnc:
